fix: widen TimeLine::time(), last_time(), last_note_time() return types from i32 to i64
Songs longer than ~35.8 minutes (2,147,483 ms) caused i64-to-i32
overflow in these methods. Change return types to i64 and propagate
the wider type through all 26 affected files across 7 crates.

- TimeLine::time(): i32 -> i64
- BMSModel::last_time(): i32 -> i64
- BMSModel::last_note_time(): i32 -> i64
- BMSPlayer::playtime and TIME_MARGIN: i32 -> i64
- AlgorithmModifyParams::now and last_note_time: i32 -> i64
- PracticeModelData::timeline_times: Vec<i32> -> Vec<i64>
- Clamp to i32 at DB/serialization boundaries (chart.length, endtime)
- Add i64 casts at mixed-type arithmetic sites (Note::time() stays i32)
- Add regression tests verifying correct values beyond i32::MAX

+++ crates/bms-model/src/bms_model_utils.rs
@@ -64,7 +64,7 @@ pub fn total_notes_full(model: &BMSModel, start: i32, end: i32, note_type: i32, 
     let lntype = model.lntype();
     let mut count = 0;
     for tl in &model.timelines {
-        if tl.time() >= start && tl.time() < end {
+        if tl.time() >= start as i64 && tl.time() < end as i64 {
             match note_type {
                 TOTALNOTES_ALL => {
                     count += tl.total_notes_with_lntype(lntype);
@@ -172,7 +172,7 @@ pub fn max_notes_per_time(model: &BMSModel, range: i32) -> f64 {
     for i in 0..tl.len() {
         let mut notes = 0;
         let mut j = i;
-        while j < tl.len() && tl[j].time() < tl[i].time() + range {
+        while j < tl.len() && tl[j].time() < tl[i].time() + range as i64 {
             notes += tl[j].total_notes_with_lntype(lntype);
             j += 1;
         }
