+++ crates/rubato-core/src/pattern/randomizer/specialized.rs
@@ -75,7 +75,9 @@ impl SRandomizer {
             if let Some(ref note) = n
                 && note.is_long()
             {
-                if note.is_end() && self.base.ln_active.contains_key(&x) && tl.time() == note.time()
+                if note.is_end()
+                    && self.base.ln_active.contains_key(&x)
+                    && tl.time() == note.time() as i64
                 {
                     self.base.ln_active.remove(&x);
                     self.base.changeable_lane.push(x);
@@ -165,7 +167,9 @@ impl SpiralRandomizer {
             if let Some(ref note) = n
                 && note.is_long()
             {
-                if note.is_end() && self.base.ln_active.contains_key(&x) && tl.time() == note.time()
+                if note.is_end()
+                    && self.base.ln_active.contains_key(&x)
+                    && tl.time() == note.time() as i64
                 {
                     self.base.ln_active.remove(&x);
                     self.base.changeable_lane.push(x);
@@ -328,7 +332,9 @@ impl AllScratchRandomizer {
             if let Some(ref note) = n
                 && note.is_long()
             {
-                if note.is_end() && self.base.ln_active.contains_key(&x) && tl.time() == note.time()
+                if note.is_end()
+                    && self.base.ln_active.contains_key(&x)
+                    && tl.time() == note.time() as i64
                 {
                     self.base.ln_active.remove(&x);
                     self.base.changeable_lane.push(x);
@@ -568,7 +574,9 @@ impl NoMurioshiRandomizer {
             if let Some(ref note) = n
                 && note.is_long()
             {
-                if note.is_end() && self.base.ln_active.contains_key(&x) && tl.time() == note.time()
+                if note.is_end()
+                    && self.base.ln_active.contains_key(&x)
+                    && tl.time() == note.time() as i64
                 {
                     self.base.ln_active.remove(&x);
                     self.base.changeable_lane.push(x);
@@ -689,7 +697,9 @@ impl ConvergeRandomizer {
             if let Some(ref note) = n
                 && note.is_long()
             {
-                if note.is_end() && self.base.ln_active.contains_key(&x) && tl.time() == note.time()
+                if note.is_end()
+                    && self.base.ln_active.contains_key(&x)
+                    && tl.time() == note.time() as i64
                 {
                     self.base.ln_active.remove(&x);
                     self.base.changeable_lane.push(x);
