+++ crates/rubato-play/src/bga/bga_processor/mod.rs
@@ -145,7 +145,7 @@ impl BGAProcessor {
             if tl.bga != -1 || tl.layer != -1 || !tl.eventlayer.is_empty() {
                 tls.push(BgaTimeline {
                     // Java TimeLine.getTime() returns (int)(time / 1000) i.e. milliseconds
-                    time_ms: tl.time() as i64,
+                    time_ms: tl.time(),
                     bga: tl.bga,
                     layer: tl.layer,
                     eventlayer: tl.eventlayer.to_vec(),
