+++ crates/rubato-play/src/bms_player/scoring.rs
@@ -358,7 +358,7 @@ impl BMSPlayer {
         self.judge.past_notes()
     }
 
-    pub fn playtime(&self) -> i32 {
+    pub fn playtime(&self) -> i64 {
         self.playtime
     }
